requirements.txt
oxforddictionaries
grpcio
grpcio-tools
protobuf